Dev para alteração de produto. Angelo PI3.

+++ LojaDeGamesPIII/src/main/java/controller/alterarProduto.java
@@ -24,8 +24,8 @@ public class alterarProduto extends HttpServlet {
         p.setId(Integer.parseInt(request.getParameter("id")));
         p.setNome(request.getParameter("nomeProduto"));
         p.setDescricao(request.getParameter("descricao"));
-        p.setPrecoDeCusto(Float.parseFloat(request.getParameter("precoCusto").replace(",", ".").replace(".", "")));
-        p.setPrecoDeVenda(Float.parseFloat(request.getParameter("precoVenda").replace(",", ".").replace(".", "")));
+        p.setPrecoDeCusto(Integer.parseInt(request.getParameter("precoCusto").replace(",", ".").replace(".", "")));
+        p.setPrecoDeVenda(Integer.parseInt(request.getParameter("precoVenda").replace(",", ".").replace(".", "")));
         p.setEstoque(Integer.parseInt(request.getParameter("estoque")));
         p.setPlataforma(request.getParameter("plataforma"));
         p.setAnoLancamento(Integer.parseInt(request.getParameter("anolancamento")));
