+++ LojaDeGamesPIII/src/main/java/controller/cadastrarProduto.java
@@ -29,8 +29,8 @@ public class cadastrarProduto extends HttpServlet {
         
         p.setNome(request.getParameter("nomeProduto"));
         p.setDescricao(request.getParameter("descricao"));
-        p.setPrecoDeCusto(Float.parseFloat(request.getParameter("precoCusto").replace(",", ".").replace(".", "")));
-        p.setPrecoDeVenda(Float.parseFloat(request.getParameter("precoVenda").replace(",", ".").replace(".", "")));
+        p.setPrecoDeCusto(Integer.parseInt(request.getParameter("precoCusto").replace(",", ".").replace(".", "")));
+        p.setPrecoDeVenda(Integer.parseInt(request.getParameter("precoVenda").replace(",", ".").replace(".", "")));
         p.setEstoque(Integer.parseInt(request.getParameter("estoque")));
         p.setPlataforma(request.getParameter("plataforma"));
         p.setAnoLancamento(Integer.parseInt(request.getParameter("anolancamento")));
